# docs/swagger.yaml
basePath: /
definitions:
  dto.DeliveryDTO:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    type: object
  dto.ItemDTO:
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        type: integer
      price:
        type: integer
      rid:
        type: string
      sale:
        type: integer
      size:
        type: string
      status:
        type: integer
      total_price:
        type: integer
      track_number:
        type: string
    type: object
  dto.OrderDTO:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/dto.DeliveryDTO'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.ItemDTO'
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        description: ID          uuid.UUID `json:"id"`
        type: string
      payment:
        $ref: '#/definitions/dto.PaymentDTO'
      shardkey:
        type: string
      sm_id:
        type: integer
      track_number:
        type: string
    type: object
  dto.PaymentDTO:
    properties:
      amount:
        type: integer
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        type: integer
      delivery_cost:
        type: integer
      goods_total:
        type: integer
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    type: object
info:
  contact: {}
  description: orders API
  title: wb_techschool 'Orders API'
  version: "1.0"
paths:
  /order/{uid}:
    get:
      description: Getting orders by UID
      parameters:
      - description: UID заказа
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderDTO'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Getting orders by UID
      tags:
      - orders
swagger: "2.0"